requirements.txt

google-adk == 1.28.0
docling == 2.82.0
sentence-transformers == 5.3.0
spacy == 3.8.14
google-genai == 1.68.0